Fix wallet services and repository integration

+++ app/models.py
@@ -1,5 +1,3 @@
-from decimal import Decimal
-
 from app.database import Base
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -9,4 +7,4 @@ class Wallet(Base):
 
     id: Mapped[int] = mapped_column(primary_key=True)
     name: Mapped[str]
-    balance: Mapped[Decimal]
+    balance: Mapped[float]
